Amounts removed from settings
