+++ support-frontend/app/admin/settings/Settings.scala
@@ -1,7 +1,5 @@
 package admin.settings
 
-import admin.settings.AmountsTests.AmountsTests
-
 import java.io.FileNotFoundException
 import java.nio.file.{Files, Paths}
 import cats.implicits._
@@ -22,7 +20,6 @@ import com.gu.support.workers.SupporterPlus
 
 case class AllSettings(
     switches: Switches,
-    amounts: AmountsTests,
     contributionTypes: ContributionTypes,
     metricUrl: MetricUrl,
     landingPageTests: List[LandingPageTest],
@@ -62,15 +59,14 @@ object Settings {
     } yield settings
 }
 
-case class SettingsSources(switches: SettingsSource, amounts: SettingsSource, contributionTypes: SettingsSource)
+case class SettingsSources(switches: SettingsSource, contributionTypes: SettingsSource)
 
 object SettingsSources {
   def fromConfig(config: Config, stage: Stage): Either[Throwable, SettingsSources] = {
     for {
       switchesSource <- SettingsSource.fromConfig(config, "switches_v2", stage)
-      amountsSource <- SettingsSource.fromConfig(config, "configured-amounts-v3", stage)
       contributionTypesSource <- SettingsSource.fromConfig(config, "contributionTypes", stage)
-    } yield SettingsSources(switchesSource, amountsSource, contributionTypesSource)
+    } yield SettingsSources(switchesSource, contributionTypesSource)
   }
 }
