+++ support-frontend/app/admin/settings/SettingsProvider.scala
@@ -17,7 +17,6 @@ import services.fastly.FastlyService
 
 import scala.concurrent.duration._
 import scala.concurrent.{ExecutionContext, Future}
-import admin.settings.AmountsTests.AmountsTests
 import services.{
   CheckoutNudgeTestService,
   LandingPageTestService,
@@ -35,7 +34,6 @@ abstract class SettingsProvider[T] {
 
 class AllSettingsProvider private (
     switchesProvider: SettingsProvider[Switches],
-    amountsProvider: SettingsProvider[AmountsTests],
     contributionTypesProvider: SettingsProvider[ContributionTypes],
     metricUrl: MetricUrl,
     landingPageTestsProvider: LandingPageTestService,
@@ -47,7 +45,6 @@ class AllSettingsProvider private (
   def getAllSettings(): AllSettings = {
     AllSettings(
       switchesProvider.settings(),
-      amountsProvider.settings(),
       contributionTypesProvider.settings(),
       metricUrl,
       landingPageTestsProvider.getTests(),
@@ -68,12 +65,10 @@ object AllSettingsProvider {
   )(implicit client: AwsS3Client, system: ActorSystem, wsClient: WSClient): Either[Throwable, AllSettingsProvider] = {
     for {
       switchesProvider <- SettingsProvider.fromAppConfig[Switches](config.settingsSources.switches, config)
-      amountsProvider <- SettingsProvider.fromAppConfig[AmountsTests](config.settingsSources.amounts, config)
       contributionTypesProvider <- SettingsProvider
         .fromAppConfig[ContributionTypes](config.settingsSources.contributionTypes, config)
     } yield new AllSettingsProvider(
       switchesProvider,
-      amountsProvider,
       contributionTypesProvider,
       config.metricUrl,
       landingPageTestService,
