+++ support-frontend/app/controllers/VatComplianceController.scala
@@ -1,13 +1,28 @@
 package controllers
 
 import actions.CustomActionBuilders
+import cats.implicits._
 import com.gu.monitoring.SafeLogging
 import play.api.Environment
-import play.api.libs.json.Json
+import play.api.libs.json.{JsValue, Json}
 import play.api.mvc.{AbstractController, Action, AnyContent, ControllerComponents}
 
-import java.nio.file.{Files, Path}
+import java.io.FileNotFoundException
+import java.nio.file.Path
+import scala.io.Source
 import scala.util.Try
+
+object VatComplianceConfig {
+
+  def fromLocalFile(path: Path): Either[Throwable, JsValue] =
+    for {
+      buf <- Try {
+        Source.fromFile(path.toFile)
+      }.toEither
+      settings <- Try(Json.parse(buf.mkString)).toEither
+      _ <- Try(buf.close()).toEither
+    } yield settings
+}
 
 class VatComplianceController(
     components: ControllerComponents,
@@ -22,18 +37,14 @@ class VatComplianceController(
     environment.rootPath.toPath.getParent.resolve("modules/VATComplianceAmounts.json")
 
   def getVatComplianceConfig(): Action[AnyContent] = CachedAction() { implicit request =>
-    if (!Files.exists(vatComplianceConfigPath)) {
-      logger.error(scrub"VAT compliance config file not found at $vatComplianceConfigPath")
-      NotFound("VAT compliance config file not found")
-    } else {
-      val rawJson = Files.readString(vatComplianceConfigPath)
-
-      Try(Json.parse(rawJson)).toEither match {
-        case Right(json) => Ok(json)
-        case Left(error) =>
-          logger.error(scrub"Failed to parse VAT compliance config JSON: ${error.getMessage}")
-          InternalServerError("Invalid VAT compliance config JSON")
-      }
+    VatComplianceConfig.fromLocalFile(vatComplianceConfigPath) match {
+      case Right(json) => Ok(json)
+      case Left(_: FileNotFoundException) =>
+        logger.error(scrub"VAT compliance config file not found at $vatComplianceConfigPath")
+        NotFound("VAT compliance config file not found")
+      case Left(error) =>
+        logger.error(scrub"Failed to load VAT compliance config JSON: ${error.getMessage}")
+        InternalServerError("Invalid VAT compliance config JSON")
     }
   }
 }
