+++ support-frontend/assets/__mocks__/settingsMock.ts
@@ -35,253 +35,6 @@ const mockSettings = {
 			enableRecaptchaFrontend: 'On',
 		},
 	},
-	amounts: [
-		{
-			testName: 'FALLBACK_AMOUNTS__GBPCountries',
-			liveTestName: '',
-			isLive: false,
-			targeting: {
-				targetingType: 'Region',
-				region: 'GBPCountries',
-			},
-			order: 0,
-			seed: 0,
-			variants: [
-				{
-					variantName: 'CONTROL',
-					defaultContributionType: 'MONTHLY',
-					displayContributionType: ['ONE_OFF', 'MONTHLY', 'ANNUAL'],
-					amountsCardData: {
-						ONE_OFF: {
-							amounts: [5, 10, 15, 20],
-							defaultAmount: 5,
-							hideChooseYourAmount: false,
-						},
-						MONTHLY: {
-							amounts: [5, 10, 15, 20],
-							defaultAmount: 20,
-							hideChooseYourAmount: false,
-						},
-						ANNUAL: {
-							amounts: [5, 10, 15, 20],
-							defaultAmount: 5,
-							hideChooseYourAmount: false,
-						},
-					},
-				},
-			],
-		},
-		{
-			testName: 'FALLBACK_AMOUNTS__UnitedStates',
-			liveTestName: '',
-			isLive: false,
-			targeting: {
-				targetingType: 'Region',
-				region: 'UnitedStates',
-			},
-			order: 0,
-			seed: 0,
-			variants: [
-				{
-					variantName: 'CONTROL',
-					defaultContributionType: 'MONTHLY',
-					displayContributionType: ['ONE_OFF', 'MONTHLY', 'ANNUAL'],
-					amountsCardData: {
-						ONE_OFF: {
-							amounts: [5, 10, 15, 20],
-							defaultAmount: 5,
-							hideChooseYourAmount: false,
-						},
-						MONTHLY: {
-							amounts: [5, 10, 15, 20],
-							defaultAmount: 5,
-							hideChooseYourAmount: false,
-						},
-						ANNUAL: {
-							amounts: [5, 10, 15, 20],
-							defaultAmount: 5,
-							hideChooseYourAmount: false,
-						},
-					},
-				},
-			],
-		},
-		{
-			testName: 'FALLBACK_AMOUNTS__EURCountries',
-			liveTestName: '',
-			isLive: false,
-			targeting: {
-				targetingType: 'Region',
-				region: 'EURCountries',
-			},
-			order: 0,
-			seed: 0,
-			variants: [
-				{
-					variantName: 'CONTROL',
-					defaultContributionType: 'MONTHLY',
-					displayContributionType: ['ONE_OFF', 'MONTHLY', 'ANNUAL'],
-					amountsCardData: {
-						ONE_OFF: {
-							amounts: [5, 10, 15, 20],
-							defaultAmount: 5,
-							hideChooseYourAmount: false,
-						},
-						MONTHLY: {
-							amounts: [5, 10, 15, 20],
-							defaultAmount: 5,
-							hideChooseYourAmount: false,
-						},
-						ANNUAL: {
-							amounts: [5, 10, 15, 20],
-							defaultAmount: 5,
-							hideChooseYourAmount: false,
-						},
-					},
-				},
-			],
-		},
-		{
-			testName: 'FALLBACK_AMOUNTS__International',
-			liveTestName: '',
-			isLive: false,
-			targeting: {
-				targetingType: 'Region',
-				region: 'International',
-			},
-			order: 0,
-			seed: 0,
-			variants: [
-				{
-					variantName: 'CONTROL',
-					defaultContributionType: 'MONTHLY',
-					displayContributionType: ['ONE_OFF', 'MONTHLY', 'ANNUAL'],
-					amountsCardData: {
-						ONE_OFF: {
-							amounts: [5, 10, 15, 20],
-							defaultAmount: 5,
-							hideChooseYourAmount: false,
-						},
-						MONTHLY: {
-							amounts: [5, 10, 15, 20],
-							defaultAmount: 5,
-							hideChooseYourAmount: false,
-						},
-						ANNUAL: {
-							amounts: [5, 10, 15, 20],
-							defaultAmount: 5,
-							hideChooseYourAmount: false,
-						},
-					},
-				},
-			],
-		},
-		{
-			testName: 'FALLBACK_AMOUNTS__Canada',
-			liveTestName: '',
-			isLive: false,
-			targeting: {
-				targetingType: 'Region',
-				region: 'Canada',
-			},
-			order: 0,
-			seed: 0,
-			variants: [
-				{
-					variantName: 'CONTROL',
-					defaultContributionType: 'MONTHLY',
-					displayContributionType: ['ONE_OFF', 'MONTHLY', 'ANNUAL'],
-					amountsCardData: {
-						ONE_OFF: {
-							amounts: [5, 10, 15, 20],
-							defaultAmount: 5,
-							hideChooseYourAmount: false,
-						},
-						MONTHLY: {
-							amounts: [5, 10, 15, 20],
-							defaultAmount: 5,
-							hideChooseYourAmount: false,
-						},
-						ANNUAL: {
-							amounts: [5, 10, 15, 20],
-							defaultAmount: 5,
-							hideChooseYourAmount: false,
-						},
-					},
-				},
-			],
-		},
-		{
-			testName: 'FALLBACK_AMOUNTS__AUDCountries',
-			liveTestName: '',
-			isLive: false,
-			targeting: {
-				targetingType: 'Region',
-				region: 'AUDCountries',
-			},
-			order: 0,
-			seed: 0,
-			variants: [
-				{
-					variantName: 'CONTROL',
-					defaultContributionType: 'MONTHLY',
-					displayContributionType: ['ONE_OFF', 'MONTHLY', 'ANNUAL'],
-					amountsCardData: {
-						ONE_OFF: {
-							amounts: [5, 10, 15, 20],
-							defaultAmount: 5,
-							hideChooseYourAmount: false,
-						},
-						MONTHLY: {
-							amounts: [5, 10, 15, 20],
-							defaultAmount: 5,
-							hideChooseYourAmount: false,
-						},
-						ANNUAL: {
-							amounts: [5, 10, 15, 20],
-							defaultAmount: 5,
-							hideChooseYourAmount: false,
-						},
-					},
-				},
-			],
-		},
-		{
-			testName: 'FALLBACK_AMOUNTS__NZDCountries',
-			liveTestName: '',
-			isLive: false,
-			targeting: {
-				targetingType: 'Region',
-				region: 'NZDCountries',
-			},
-			order: 0,
-			seed: 0,
-			variants: [
-				{
-					variantName: 'CONTROL',
-					defaultContributionType: 'MONTHLY',
-					displayContributionType: ['ONE_OFF', 'MONTHLY', 'ANNUAL'],
-					amountsCardData: {
-						ONE_OFF: {
-							amounts: [5, 10, 15, 20],
-							defaultAmount: 5,
-							hideChooseYourAmount: false,
-						},
-						MONTHLY: {
-							amounts: [5, 10, 15, 20],
-							defaultAmount: 5,
-							hideChooseYourAmount: false,
-						},
-						ANNUAL: {
-							amounts: [5, 10, 15, 20],
-							defaultAmount: 5,
-							hideChooseYourAmount: false,
-						},
-					},
-				},
-			],
-		},
-	],
 	contributionTypes: {
 		GBPCountries: [
 			{ contributionType: 'ONE_OFF' },
